HSA version V2.2b
1.NTB bugs fixed.

 Changes to be committed:
	modified:   src/main/java/net/minecraft/server/BlockPiston.java
	modified:   src/main/java/net/minecraft/server/ChunkProviderServer.java
	modified:   src/main/java/net/minecraft/server/ChunkRegionLoader.java
	modified:   src/main/java/net/minecraft/server/MinecraftServer.java
	modified:   src/main/java/org/bukkit/craftbukkit/CraftChunk.java
	modified:   src/main/java/org/spigotmc/AntiXray.java

+++ src/main/java/org/bukkit/craftbukkit/CraftChunk.java
@@ -300,7 +300,7 @@ public class CraftChunk implements Chunk {
                 hebiomeTemp = new double[256];
                 hebiomeRain = new double[256];
                 //float[] dat = getTemperatures(wcm, x << 4, z << 4);
-                hedat = getTemperatures(wcm, (int)(x << 4), (int)(z << 4));
+                hedat = getTemperatures(wcm, x << 4, z << 4);
                 
                 Aparapi.range(256).forEach(gid_i -> {
                     hebiomeTemp[gid_i] = hedat[gid_i];
